Scripts files refactored
