+++ events/blocks/events-form/events-form.js
@@ -1,10 +1,10 @@
 import { LIBS } from '../../scripts/scripts.js';
-import { getMetadata } from '../../utils/utils.js';
+import { getMetadata } from '../../scripts/utils.js';
 import HtmlSanitizer from '../../deps/html-sanitizer.js';
-import { createAttendee, deleteAttendee, updateAttendee } from '../../utils/esp-controller.js';
+import { createAttendee, deleteAttendee, updateAttendee } from '../../scripts/esp-controller.js';
 import BlockMediator from '../../deps/block-mediator.min.js';
 
-const { createTag } = await import(`${LIBS}/utils/utils.js`);
+const { createTag } = await import(`${LIBS}/scripts/utils.js`);
 const { closeModal } = await import(`${LIBS}/blocks/modal/modal.js`);
 const { default: sanitizeComment } = await import(`${LIBS}/utils/sanitizeComment.js`);
