+++ events/blocks/preview-bar/preview-bar.js
@@ -1,7 +1,7 @@
 import { LIBS } from '../../scripts/scripts.js';
-import { getIcon } from '../../utils/utils.js';
+import { getIcon } from '../../scripts/utils.js';
 
-const { createTag, getMetadata } = await import(`${LIBS}/utils/utils.js`);
+const { createTag, getMetadata } = await import(`${LIBS}/scripts/utils.js`);
 
 function removeURLParameter(url, parameter) {
   const params = new URLSearchParams(url.search);
