+++ events/scripts/utils.js
